+++ src/_scaffolding/services/bb_rest_services.go
@@ -1,0 +1,6 @@
+package services
+
+/**
+ * The [BbRestServices] interface...
+ */
+type BbRestServices interface {}